fix: 설문 점수 소수 저장 적용

+++ src/main/java/com/capteam/gaobackend/dto/user/request/UserSurveyRequestDto.java
@@ -69,19 +69,19 @@ public class UserSurveyRequestDto {
     public static class PersonalityScoresDto {
 
         // 소통 성향 점수를 저장하는 필드입니다.
-        private Integer communication;
+        private Double communication;
 
         // 책임감 성향 점수를 저장하는 필드입니다.
-        private Integer responsibility;
+        private Double responsibility;
 
         // 협업 성향 점수를 저장하는 필드입니다.
-        private Integer collaboration;
+        private Double collaboration;
 
         // 유연성 성향 점수를 저장하는 필드입니다.
-        private Integer flexibility;
+        private Double flexibility;
 
         // 감정 안정성 점수를 저장하는 필드입니다.
-        private Integer emotionalStability;
+        private Double emotionalStability;
     }
 
     @Getter
@@ -89,18 +89,18 @@ public class UserSurveyRequestDto {
     public static class DevelopmentScoresDto {
 
         // 리더십 성향 점수를 저장하는 필드입니다.
-        private Integer leadership;
+        private Double leadership;
 
         // 문제 해결력 점수를 저장하는 필드입니다.
-        private Integer problemSolving;
+        private Double problemSolving;
 
         // 구현 실행력 점수를 저장하는 필드입니다.
-        private Integer implementation;
+        private Double implementation;
 
         // 학습 성장성 점수를 저장하는 필드입니다.
-        private Integer learningAbility;
+        private Double learningAbility;
 
         // 기획 정리력 점수를 저장하는 필드입니다.
-        private Integer planning;
+        private Double planning;
     }
 }
